+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Please, check out guidelines: https://keepachangelog.com/en/1.0.0/
 
 ## Next version
 
+## 0.4.0
+
 ### Added
 
 - Throw an error if a library target contains resources https://github.com/tuist/tuist/pull/98 by @pepibumur.